+++ src/revolut/client.py
@@ -10,25 +10,21 @@ from ._http.sync_client import SyncTransport
 from .config import DEFAULT_API_VERSION, DEFAULT_TIMEOUT, Environment
 from .resources import (
     AsyncCustomersResource,
-    AsyncDisputesResource,
     AsyncLocationsResource,
     AsyncOrdersResource,
     AsyncPaymentMethodsResource,
     AsyncPaymentsResource,
     AsyncPayoutsResource,
     AsyncRefundsResource,
-    AsyncReportRunsResource,
     AsyncSubscriptionsResource,
     AsyncWebhooksResource,
     CustomersResource,
-    DisputesResource,
     LocationsResource,
     OrdersResource,
     PaymentMethodsResource,
     PaymentsResource,
     PayoutsResource,
     RefundsResource,
-    ReportRunsResource,
     SubscriptionsResource,
     WebhooksResource,
 )
@@ -67,8 +63,6 @@ class RevolutMerchantClient:
         self.payment_methods = PaymentMethodsResource(self._transport)
         self.subscriptions = SubscriptionsResource(self._transport)
         self.payouts = PayoutsResource(self._transport)
-        self.disputes = DisputesResource(self._transport)
-        self.report_runs = ReportRunsResource(self._transport)
         self.locations = LocationsResource(self._transport)
         self.webhooks = WebhooksResource(self._transport)
 
@@ -115,8 +109,6 @@ class AsyncRevolutMerchantClient:
         self.payment_methods = AsyncPaymentMethodsResource(self._transport)
         self.subscriptions = AsyncSubscriptionsResource(self._transport)
         self.payouts = AsyncPayoutsResource(self._transport)
-        self.disputes = AsyncDisputesResource(self._transport)
-        self.report_runs = AsyncReportRunsResource(self._transport)
         self.locations = AsyncLocationsResource(self._transport)
         self.webhooks = AsyncWebhooksResource(self._transport)